dev/kaos_dev_ig_stream_input/src/requirements.txt
git+https://github.com/ig-python/ig-markets-api-python-library
munch==2.5.0
pycryptodome==3.9.8
requests==2.24.0
requests-cache==0.5.2
six==1.15.0
pandas==1.2.0 ; python_version >= '3.7.1'
pandas==1.1.5 ; python_version < '3.7.1'
google-cloud-pubsub
google-auth
